LPS-179353 Extract ScreenNavigationEntry methods from CPDefinitionSubscriptionScreenNavigationCategory to CPDefinitionSubscriptionScreenNavigationEntry, extends CPDefinitionSubscriptionScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
